agent/requirements.txt
fastmcp==2.10.1
requests>=2.28.0
aiohttp>=3.8.0
openai==1.93.0